Move syntaxhighlighters to utils namespace
Two reasons motivate this change:

1. This file was in widgets/sourcecode but it doesn't contain any widget
2. It will let us avoid circular imports when adding things to
widgets/__init__.py
