+++ spyderlib/plugins/outlineexplorer.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2011 Pierre Raybaut
+# Copyright © 2009- The Spyder Development Team
 # Licensed under the terms of the MIT License
 # (see spyderlib/__init__.py for details)
 
@@ -8,7 +8,7 @@
 
 Data for outline are provided by method .get_outlineexplorer_data() of
 highlighter of assigned editor. For example, for Python files code editor uses
-highlighter spyderlib.widgets.sourcecode.syntaxhighlighters.PythonSH
+highlighter spyderlib.utils.syntaxhighlighters.PythonSH
 """
 
 from spyderlib.qt.QtCore import Signal
